Include service manual in unified_search searches
Note: depends on
https://github.com/alphagov/government-service-design-manual/pull/545
being deployed first, but I'll deploy that in the same deploy slot, so
please merge when ready.

Now that the service manual is (or will be) sending documents to
rummager in the correct format for manuals, include the `service-manual`
index when searching using the `unified_search.json` endpoint.

This will allow the service-manual app to start using the unified_search
endpoint, and thus allow us to deprecate the now-unused old endpoints.

This also allows the service manual and its sections to come up in
search. In order to make sure that they don't come up too high for
unrelated searches, this commit adjusts the weight of anything from the
`service-manual` index downwards considerably. Once the documents are
available in search, we can review whether this down-weighting is
required, or excessive.