Downweight historic WH content
We want to downweight historic content so users are less likely
to be shown historic content when using a broad search term. When
searching specifically for historic content it should still rank
highly. Eg, when searching for something like "tax" historic
content should be less visible, but when searching for the title
of a speciic historic tax document it should still be top.

(Historic means that the content is political and published under
a prior government, eg, not relflecting the political stance of
the current government)

Non political content will be unaffected.

50% downweight is an initial weighting, based on being less visible
than before, but more visible than closed/devolved orgs.

This value will be updated as we get more analytics/insights.