Add CombinedIndexSchema
In order to replace parsing of elasticsearch mappings in
lib/search_parameter_parser.rb, we need to be able to get a list of the
field definitions (and their allowed_values lists) for all document
types in all the indexes being searches.  The allowed_values will be
looked up to return human readable names for some of the fields.

To do this, introduce a `CombinedIndexSchema` object;
`CombinedIndexSchema.field_definitions` returns a hash from field_name to a
`FieldDefinition` object, for all fields allowed in any document type in
the index.

This is more complicated than I would like, because FieldDefinitions
includes "allowed_values", which can differ for a single field between
document types.  In this case, we merge the allowed values.  Future work
may seek to define allowed_values on the field definition, rather than
the document type, which would avoid the complexity of having to be able
to merge field definitions.