Short-term hack to treat public_timestamp as a date
This will force `public_timstamp` to be a date-field for filtering.
Without a `document_type` a schema can't be matched (see `schema`) and a
type can't be found, defaulting to string, and causing an elastic search
error.

Date filtering is requried for the new policy finders, which will need it
as part of the election work, so is requried quite soon.

There is larger refactoring work inprogress by @rboulton to fix this
properly, but for now this unblocks the election team for now. It's
expected the larger refactor work will be done before the election, but
this short term hack will allow the election team to test this on
preview, and build other features on top of it, in the mean time.