Add health check for spelling suggestions
We are currently working on improving the spelling suggestions for
search. This commit adds health checks for suggestions.

It works by downloading a list of search terms and their expected
results from a public Google Doc. The terms are then compared to the
actual suggestions returned from gov.uk search.

The new health check reports the percentage of correctly corrected
search terms.