Add a `spelling_text` field for suggestions
This field contains a copy of all the text that's used for searching,
tokenised only by splitting it into words and groups of words
("shingles") and lowercasing.

Stemming, stopwording, etc, are not applied. This is to support
spelling correction, which otherwise will try to correct words to their
stemmed forms.  eg "prisioner" would be corrected to "prison" if
corrected using the `_all` field.  Using the `spelling_text` field, it
is corrected to "prisoner".

Lowercasing is needed so that words which are consistently capitalised
in our documents (such as "PAYE"), but are not consistently capitalised
in searches, are still corrected to useful variants.  It would be nice
to avoid having to lowercase, so that we could, for example, suggest
proper nouns with the appropriate capitalisation, but that is less
important than making basic corrections work at present.

The shingles filter adds "shingles" (token n-grams) to the field to
allow corrections like "driving liecence" to "driving licence". From
the docs:

> 'For example, the sentence "please divide this sentence into
shingles" might be tokenized into shingles "please divide", "divide
this", "this sentence", "sentence into", and "into shingles".'

The field is generated by elasticsearch at index time by copying any
`searchable_text` or `searchable_identifier` fields.  It therefore
won't appear in the _source field when fetching documents, or be shown.

However, it is possible to search on it, and use it as the target field
for a TermSuggester for spelling corrections, etc.

Original commit by @rboulton