Add filter_type configuration for field types
When filtering on a field, we need to build filters in different ways
depending on the type of the field.

Currently, we define two filter types: "text" for textual fields, and
"date" for date fields.  (These are implemented by TextFieldFilter and
DateFieldFilter in search_parameter_parser.)

Add a "filter_type" setting which can be set on a field type (in
`field_types.json`) to indicate how to build filters for that field.  If
omitted, filtering is not allowde on that field.

Add an `allowed_filter_fields` method to CombinedIndexSchema which
returns the list of fields in the index for which a `filter_type` is
defined.  This will be used by SearchParameterParser in a later commit.