Simplify checking fields in parameter parser
Remove the hardcoded list of "ALLOWED_FILTER_FIELDS" in favour of
allowing any field which has a defined filter type to be used for
filtering.  This increases the number of fields which can be filtered on
considerably, but I'm happy with this since filtering imposes little
overhead in elasticsearch, so if callers want to do that we should let
them.

It also removes the restriction that fields which are defined for only a
few document types (such as finders) can only be filtered on if a
document_type restriction is added.  This was only ever done because it
was the easiest thing to implement, and removing it makes the code much
cleaner and easier to document.

"document_type" needs to be added as a special case for the list of
allowed filter fields, since it's actually an alias for the internal
"_type" field, and therefore doesn't appear in the schema.

As part of this, SearchParameterParser now needs to be given a
CombinedIndexSchema for the indexes which are being searched on;
previously, it was being given an `Index` object for the 'mainstream'
index, which was used to get the schema for the "finder" document types,
for allowing extra filter fields when filtering by `document_type`.