Add elasticsearch-based spell check
Adds spell checking using elasticsearch's Phrase Suggester feature.
Suggestions made by elasticsearch are more accurate, mainly because the
suggestions are drawn from the content on GOV.UK itself.

- When user searches we now do an extra request to elasticsearch to
fetch suggestions for the search term. This has to be done in a
separate request because of limitations in working with multiple
indices (more in `spell_check_fetcher.rb`).

- The "suggest" part of the elasticsearch query is built in
`QueryComponent::Suggest`. The pattern of using small builder classes
to build the query should be used to refactor `UnifiedSearchBuilder`.

- `SuggestionBlacklist` makes sure we don't suggest corrections for
search terms we deem correct.