Introduce all_searchable_text field
Elasticsearch's built in `_all` field is useful to be able to search
across all fields (or at least, all for which `include_in_all` is true).
However, it's rather special and inflexible; in particular, it's not
possible to use the "multi fields" option to index it in multiple ways.

To allow this, and in particular to allow searching across all text both
with and without synonym expansion, this commit introduces a custom
`all_searchable_text` field, which is populated using the `copy_to`
directive for any searchable field (ie, any for which `include_in_all`
is currently true).  This field is indexed both by the standard
analyzer, and to a subfield with synonym expansion.

Ultimately, this will allow us to disable the "_all" field entirely, but
this isn't yet possible because old endpoints continue to use it, and
also the existing endpoint will use it until we move over to the new
weighting scheme.