Ignore names in spelling suggester
Because the names of new cabinet members do not occur a lot yet in our
documents, we will have to temporarily ignore them manually to prevent
potentially awkward spelling corrections.

These can be removed in the future when elasticsearch has "learned"
that their names are correct because they are found often in the
government/mainstream indices.