Expand 'people' fields into person objects
Results with a 'people' field were returning an array of string
slugs, compared to fields like 'organisation' where they are
expanded into full objects, with the title/link/etc of the
organisation included.

This adds the appropriate code to expand people fields, so that
the full object representation can be used to display details
about people, eg, in metadata or filter facets in a Finder.