Refactor UnifiedSearchBuilder
UnifiedSearchBuilder is responsible for taking a set of parameters
(like search string, filters and facets) and producing a search request
body that is sent to elasticsearch.

This PR refactors UnifiedSearchBuilder to improve readability and
maintainability. This is done by splitting it up into QueryComponents
that are responsible for a specific piece of the search query.

There are two types of components:

- Components that take the params hash and return a payload for use in
the query. For example, the QueryComponents::Sort returns the correct
sort clause for a set of params.
- Components that #wrap a query to "enhance" it. For example,
QueryComponents::Booster takes an existing query and returns a query
that uses booststo promote certain types of content.
The tests for UnifiedSearchBuilder have also been split up and
refactored.