Make title ordering case and whitespace insensitive
Instead of not analyzing the title.sort field at all, strip leading and
trailing whitespace and lowercase it.  This will make the sort case
insensitive, and not dependent on whitespace, which is what is usually
meant when people sort by alphabetical order.

Also, add an integration test for this sorting.

This commit introduces a new filter for the processing we want for
sorting: `string_for_sorting`.  This is actually identical currently to
the existing `exact_match` analyzer, but it's good to have analyzers
named by function, so that we can improve them in future without having
to work out what things are using them.

Also, in order to test the case insensitivity, downcase the word
"Sample" in the title of some of the sample documents.