Normalise apostrophes in spelling suggestions
Elasticsearch sees words with a "normal" apostrophe as one word, but
not words with a curly apostrophe. Before indexing the search
suggestions, we normalise all variations on the single quote to the
normal apostrophe: U+0027.

This prevents words like "jobseeker's allowance" (normal apostrophe) to
be corrected to "jobseeker's allowance" (curly apostrophe).