Preserve the "_id" field when migrating an index
Currently, when an index is migrated, Document objects are created from
the _source object for each document in the index.  Unfortunately, the
_source object doesn't contain the "_id" field, which the Document
constructor looks for to assign an _id field.  For most indexes, this
doesn't matter because if the Document constructor can't find an "_id"
field, it uses the "link" field instead.  However, for the
"page_traffic" and "metasearch" indexes, there isn't a link field.

This means currently that migrating these indexes changes all the
document IDs to ones randomly assigned by Elasticsearch.  This commit
changes this so that document IDs are preserved when migrating these
indexes.