Normalize and strip apostrophes when indexing/searching
Currently, words with curly apostrophes (copy-pasted from a word
document for example) are tokenised differently from words with
"normal" apostrophes. This causes different documents to turn up
depending on the type of apostrophe is used.

This commit adds character filters to the analysers that first
normalise the apostrophe and then remove it altogether in certain
situations.