Use shingles properly
In the existing weighting algorithm, we attempt to use "shingles" to
boost the score for documents where some of the words in the query occur
as a phrase (but maybe not all of them).

Shingles are an elasticsearch feature which involves taking the terms
generated by an analyzer, and indexing pairs of them to represent words
which occur together, instead of indexing them directly.  (It can also
produce triples, etc, but we have it configured to only produce pairs.)

The existing weighting only applies this analysis at search time, so the
pairs it looks for never exist in the index.  This commit instead
creates a new pseudo-subfield for searchable text fields which is
indexed and searched with shingles applied.  The weights from searches
across this field are used to boost scores of documents.

This is relatively expensive in terms of space; it
increases the index size for the government index by about 30%.
However, our indexes are pretty small at present (around 1.5Gb total),
so this is easy to afford.