Reduce effect of popularity for very popular pages
Currently, the popularity ranking formula is to multiply the score by:

  (1 / rank_of_document) + 0.001

What we're actually aiming to do is to roughly multiply the score of a
page by the amount of traffic it gets.  However, if we used the actual
figure for the amount of traffic directly, it would be quite unstable;
all the rankings could vary wildly if there was a particular event
happening causing some pages to get much more traffic than normal.

Instead of using the figures directly, therefore, we use the inverse
rank of the page.  Ie, the top page has a rank of 1, the next page has a
rank of 2, etc.  This is a good (but relatively robust to traffic
bursts) approximation to the traffic a page gets, because the amount of
traffic pages on our site get follows a Zipf distribution
(http://en.wikipedia.org/wiki/Zipf%27s_law).

The "0.001" part of this has two purposes.  Firstly, for brand-new pages
there is no "rank", so we give that part of the formula a value of 0.
This would mean that brand-new pages always got a score of 0, so we add
a small constant to formula.  Secondly, for existing but rarely visited
pages, the exact rank becomes a sensitive to small fluctuations in
traffic, so adding a small constant "buffers" this effect.

This formula has worked reasonably well for some time, but we often see
high traffic pages swamping the results a bit too much.  For example,
the HMRC page (currently at rank 3) is often ranked highly for tax
related searches, over the more-specific content that matches the
search.  This commit adds a mitigation for high traffic pages, by adding
a constant to the `rank_of_document` part of the formula.  To start
with, I'm setting this constant to `10`, which means, eg, that instead of the
page at rank 3 getting 30 times the multipler of something at rank 100,
it would only get around 7.7 times the multiplier.  For lower ranked
documents, this adjustment will have very little effect.

Experimenting locally, this seems to improve quite a few searches, but
still return high-ranked pages highly.