Update readme
Rummager's Readme hasn't been reviewed for a long time. This PR updates
the Readme to reflect changes and to make it conform to the styleguide.

The readme sections now conform to the styleguide.
Non-essential information on the popularity scoring and health checks
has been moved to the /docs directory to keep the Readme short.
Information on the size of the dev VM has been removed. Maybe this
belongs more in the docs for the VM itself.
Adds an appropriate license file