Add detailed_format to edition
As we need to get back a slug version of `display_type` I added a new
field to edition and gave it all the possible values of display type to
the `allowed_values` field in the edition doctype.

The need for this comes from the fact that `display_type` is sent to
rummager as a title case string `News story` instead of `news-story`.

Ticket:
https://trello.com/c/TAuAS4wp/174-8-add-more-filter-options-to-policy-pages