Parallelise bulk loading of a new index.
We can get much more performance while bulk loading a new index (either
migrating an indes, or loading an index from an external dump), by
processing batches of records in parallel.  This is because it allows
elastisearch to use all the resources available to it.  See [1] for more
details.

I've picked a default concurrency level of 12 to match our
infrastructure.  Elasticsearch defaults to using an indexing thread pool
size that matches the number of CPU cores available.  Our
api-elasticsearch machines have 4 cores, and we run a 3-node cluster,
hence a value of 12 makes sense. This may need to be tuned after
benchmarking etc.

This change halved the indexing time on my dev VM.

[1] https://www.elastic.co/blog/performance-considerations-elasticsearch-indexing