Generate snippet for search results
Currently, the frontend application does a lot of work to generate
descriptions for the search results. For example by adding missing
descriptions for topic pages. This commit adds a Snippet class which
replicates the current behaviour of frontend.

Having the behaviour here will allow us to add highlighting, show
fragments (parts of the content) and to serve the same snippet to other
users of rummager.