Remove CLI search
There's currently a `bin/search` file which could be used to get data
on search results.

Originally it used pretty much the same code as normal search, but that
has diverged significantly in the last few years. This means it doesn't
actually return the same search results as the website does, making it
less useful. And it has probably not been used for a while.