Add /content endpoints
The /content endpoint allows the user to fetch a document with a
certain link and get back the raw elasticsearch document.

For example: GET /content?link=/vehicle-tax Will return a JSON hash
with the `raw_response`.

The DELETE /content?link=/vehicle-tax endpoint will remove the document
from the search results.