Temporarily fix leading slashes in links
The link field should contain either a full URL or a relative URL with
leading slash for the frontend to work. Manuals and specialist
documents currently send links without leading slash to rummager,
causing problems in search and collections.

This temporarily adds the leading slash. This functionality should be
removed once all documents are indexed correctly. For this reason the
tests have been isolated in a separate file.