Only amend search description when necessary
We currently amend the search description for organisations to make the
text for organisations read nicely.

For example:

> Department for Environment, Food & Rural Affairs
> The home of Department for Environment, Food & Rural Affairs on
GOV.UK. We are the UK government department responsible for policy and
regulations on environmental, food and rural issues.

We'd like Whitehall to add the "Home of X on GOV.UK", not rummager.
This commit adds a guard to displaying the prefix, so that when the
changes on whitehall are deployed, we don't show the prefix twice. The
entire logic can be deleted when we've deployed and fixed whitehall.