Add script to cleanup old sitemaps
Currently we are keeping all sitemaps ever created. It would be good to
clean then up after a couple of days so that we don't have hundreds of
them saved on disk.

This lets you set a number of days to clean up and then deletes all
sitemaps which aren't the number of days most recent. If the sitemaps
are not created on a particular day this will still maintain at least
the right number of days worth of sitemaps not just the last number of
calendar days worth.