Remove `snippet` field
The `snippet` field way meant to be a transitional field in the
migration from descriptions modified by frontend to a world where the
description is indexed in elasticsearch. It is not used anywhere on the
frontend and can be safely removed.

This will allow for a cleaner PR for the highlighting.