Add highlighted title and description
This commit allows users to request two virtual fields in the search
results: `title_with_highlighting` and `description_with_highlighting`.

These contain the title and description with `<em>` tags around the
words the user was looking for. Elasticsearch automatically highlights
stemmed words and synonyms, so when the user searches for "oversea",
the words "overseas" and "abroad" will also be highlighted.

About the configuration of the highlight query: `number_of_fragments`
specifies how many "fragments" elasticsearch will return. We set
`number_of_fragments: 0` for titles to make ES return the entire
string. For descriptions we use `number_of_fragments: 1` to enable
truncation. This is preferable to doing the truncation ourselves,
because ES can do truncation on both sides of the description by
"centering" on the highlighted word. This does mean we need some voodoo
to append/prepend ellipses the text.