Stop using `Document` for Registries
`Index#documents_by_format` and `IndexForSearch#documents_by_format`
use the Document class to wrap results. This is not necessary - the
only task it performs is converting the elasticsearch response, which
return all fields as arrays.

We add a MultivalueConverter class to perform this task.

This cleans up the tests and removes the use of the `Document` from the
search-side.