Only enable highlighting when requested
Currently we request highlighting from elasticsearch for all queries,
even when the client hasn't asked for either `title_with_highlighting`
or `description_with_highlighting`. This means elasticsearch has to do
work that's not needed.

This commit makes it so that we highlight if the client requests
_either_ title or description. We could make it so that it doesn't
highlight description when only asking for the title and not
description, but that would require more code and is a use case not
likely to happen.