Remove section field from tests
Many of the tests assumed that a "section" field existed.  Now that this
is removed, update the tests to use the "mainstream_browse_pages" field
instead.

In many places this is a simple mechanical substitution, but I've tried
to use slightly plausible values for the field where this is easy.  In
other cases the "section" field was being put into stubs, but not ever
referenced by tests, so I've removed it.