Remove tags field
The tags field isn't used anywhere: it was an aborted attempt to unify
all the tagging information into a single elasticsearch field, so that a
combined facet across all of the tags could be calculated.  We're not
actually doing this, so currently the field isn't used at all and might
as well be removed.