Only check spelling when explicitly requested
We currently do spelling corrections for all search requests, even when
the client hasn't asked for it and won't use it. Since the suggester
makes a separate HTTP query to elasticsearch this has significant
overhead.

We've previously introduced the `suggest=spelling` parameter
(https://github.com/alphagov/rummager/pull/511) and updated the sole
user of the feature, frontend, to send that parameter
(https://github.com/alphagov/frontend/pull/880).

We can now restrict suggesting to clients that request it.

There are no new tests because the behaviour hasn't changed - the
existing acceptance test already has the `suggest=spelling` parameter,
so it will keep working. I've resisted the urge to test the negative
case, because that feels like overkill (testing the absence of things
not asked seems potentially endless) - but I'm open to people's
thoughts on the matter.