Simplify test helpers and insert data into search directly
This PR simplifies the test helpers by letting all integration tests
inherit from the same `IntegrationTest` class and inlining methods and
modules. This makes the code more easy to follow as we don't have to
think about the inheritance of `setup` and `teardown` blocks, which
have been inlined.

All this enables us to use consistently insert testing data directly
into elasticsearch. Currently this is done via the app itself (by
calling `post "/documents"` for example). This tightly couples the
indexing of the application to the search. It caused particular pain in
the implementation of https://trello.com/c/xQtX24kW, which is now
significantly easier.

It would've been better if this was two PRs, but it was hard to
separate the two goals here. I can walk provide more explanation of the
refactors if needed.