Remove redundant test for document deletion
This test checks if when we delete a document, the document won't
appear in the search index. But the test above that tests that
documents are correctly deleted, so this test doesn't test anything
new. In the interest of an easier test code base we should remove tests
that don't add value.