Rewrite BestBetsTest
This commit rewrite the BestBetsTest so that is doesn't depend on the
global sample data to be present. By adding documents and bets right in
the test method we make the tests self-contained and readable. Also
speeds up the runtime from ~10s to ~4s because we don't have to insert
that much data per test.