Allow sorting by tribunal decision date
HMCTS would like the search results for
tribunal decisions to be sorted by the
decision date, with the most recent decisions
listed first.

The same field "tribunal_decision_decision_date"
is present in each of the tribunal finder
configurations. So this addition to the
ALLOWED_SORT_FIELDS hash only needs to happen
once.