Refactor the way facets are filtered
Facets return a `count` with the number of documents in the index that
match that facet. This is primarily used to display the number of
documents-per-organisation in the organisation select box on the search
page, for example "HM Revenue & Customs (5,175)" and "Cabinet Office
(440)".

In most cases, when we are filtering via a facet, we don't want to
apply this filter to the list of facets. That would mean that if we
filtered on HMRC, the Cabinet Office entry would show "(0)" documents,
because there obviously aren't any Cabinet Office documents in the
results for HMRC. That's why we don't filter facets on the chosen
filter.

This commit moves the code to decide which filters to show up to the
`facets`, so that the logic described above is clearer and in one place.