Refactor TagLookup
This commit refactors the `TagLookup` class.

The immediate motivation for this is that we tried understanding the
transformation process during an incident-related investigation. Usage
of `merge`, `concat` and `<<=` all made it hard to immediately grasp
what was going on.

This commit replaces the core logic of the class (adding tags to the
doc_hash) by doing everything in one method and duplicating code. This
feels really odd, because DRY, but I think the result is much more
readable.