Implement Exact Phrase Matching
If the search term is entirely enclosed in double quotes, this is an exact phrase match.
The search term "London Transport" will return documents that have the phrase "London Transport" in them, but not those that have "London" and "Transport" with other words in between.
Synonyms are not used when doing an exact phrase match.  Therefore searching for "Driving abroad" will not return documents that have "Driving overseas" in them.
Stemming is NOT turned off: so "London Transport" will return "London's Transport"